move getVariableSignature to execution
as it cannot be used by validation, which must collect all errors rather than fail with invalid type for signature
